// web/src/api/generated/AddressTxsPage.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TxDetailDto } from "./TxDetailDto";

/**
 * `GET /address/{addr}/txs`.
 */
export type AddressTxsPage = { address: string, items: Array<TxDetailDto>, total: number, next_cursor: string | null, };